+++ src/ssdp/SSDPDB_ObjC.mm
@@ -135,8 +135,8 @@ private:
 }
 
 
--(int)addObserver:(SSDPDB_ObjC_Observer*)obs{
-    int ret = 0;
+-(NSUInteger)addObserver:(SSDPDB_ObjC_Observer*)obs{
+    NSUInteger ret = 0;
     [self lock];
     [mObservers addObject:obs];
     ret = [mObservers count];
@@ -144,8 +144,8 @@ private:
     return ret;
 }
 
--(int)removeObserver:(SSDPDB_ObjC_Observer*)obs{
-    int ret = 0;
+-(NSUInteger)removeObserver:(SSDPDB_ObjC_Observer*)obs{
+    NSUInteger ret = 0;
     [self lock];
     [mObservers removeObject:obs];
     ret = [mObservers count];
